\name{.merge.lists}
\alias{.merge.lists}
\title{Merging Lists}
\usage{
  .merge.lists(a, b)
}
\arguments{
  \item{a}{the first list}

  \item{b}{the second list}
}
\value{
  the merged list
}
\description{
  \code{.merge.lists} takes two lists and merges them.
}
\keyword{internal}

